Add initial BMP390 and u-blox driver changes

+++ Core/Inc/bmp390_runner.h
@@ -27,8 +27,8 @@
 
 /** Public variables. *********************************************************/
 
-extern double bmp390_temperature;
-extern double bmp390_pressure;
+extern float bmp390_temperature;
+extern float bmp390_pressure;
 
 /** Public functions. *********************************************************/
